+++ cli/src/commands/git/push.rs
@@ -553,6 +553,7 @@ fn find_branches_targeted_by_revisions<'a>(
         let current_branches_expression = RevsetExpression::remote_branches(
             StringPattern::everything(),
             StringPattern::exact(remote_name),
+            None,
         )
         .range(&RevsetExpression::commit(wc_commit_id))
         .intersection(&RevsetExpression::branches(StringPattern::everything()));